Agregando posibilidad de definir detail (formulario hijo) y is_detail (en el hijo, para poner un formulario detalle -hijo) en el tables, componiendo error en campo auto
git-svn-id: http://svn.almidon.org/trunk@186 924a4236-293c-0410-b38f-c0813de1d4fd

+++ php/404.php
@@ -42,6 +42,12 @@ if ($object) {
   }
   $ot = $object . 'Table';
   $$object = new $ot;
+  #If I'm a detail, not the master
+  if($$object->is_detail) {
+    require(ALMIDONDIR . '/php/detail.php');
+    die();
+  }
+  # If it continues is because I'm master
   require(ALMIDONDIR . '/php/typical.php');
   $$object->destroy();
   $tpl = ($$object->cols > 5) ? 'abajo' : 'normal';
